+++ example/lib/pages/home/home_page.dart
@@ -10,6 +10,7 @@ import 'package:example/pages/input/mouse/mouse_input_page.dart';
 import 'package:example/pages/input/move_camera_mouse/move_camera_page.dart';
 import 'package:example/pages/input/tap/tap_gesture_page.dart';
 import 'package:example/pages/lighting/lighting_page.dart';
+import 'package:example/pages/map/infinite/infinite_map_page.dart';
 import 'package:example/pages/map/spritefusion/spritefusion_page.dart';
 import 'package:example/pages/map/terrain_builder/terrain_builder_page.dart';
 import 'package:example/pages/map/tiled/tiled_network_page.dart';
@@ -139,6 +140,12 @@ class _HomePageState extends State<HomePage> {
             builder: (_) => const TerrainBuilderPage(),
             codeUrl:
                 'https://github.com/RafaelBarbosatec/bonfire/blob/develop/example/lib/pages/map/terrain_builder',
+          ),
+          ItemDrawer(
+            name: 'Infinite world (Tiled)',
+            builder: (_) => const InfiniteMapPage(),
+            codeUrl:
+                'https://github.com/RafaelBarbosatec/bonfire/blob/develop/example/lib/pages/map/infinite',
           ),
         ],
       ),